pemograman-kripto/05-pertemuan-5/pemograman-kripto/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
asyncpg>=0.29.0
python-dotenv>=1.0.1
